backend/requirements/base.txt
# Django et Django REST Framework
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
python-dotenv==1.0.1

# PostgreSQL
psycopg2-binary==2.9.9

# Utilitaires
Pillow==10.1.0
numpy==1.26.3
pandas==2.1.4

# NLP pour l'analyse de sentiment
nltk==3.8.1
textblob==0.17.1
scikit-learn==1.3.2